Feat multisample (#4)
* `SampleOptions` now supports multiple sampling modes for more control over sampling without using the Candidate API directly.
* Update `dittomancer` for new API changes

We managed to make it uglier. The example should be rewritten, but it works.

* fix `regurgitater` example

We made the API uglier. We do need an iterator that yields pieces *and* tokens. The previous API was nice like that.

It is possible to return to the old API, since yet another iterator wrapper should be trivial.