update `llama.cpp`
The candidate crash was not in fact caused by bad data from the kernel or ids out of range, rather my fault. So nothing to do with this issue here:

https://github.com/ggerganov/llama.cpp/issues/7476

The `Candidates` container can shrink, but it's only used in locally typical sampling, and I forgot it did that which is why ids are out of range in that case. I could use a HashMap to solve i or an associative array, but it's easier just to remove the duplicate check from `Candidates::from_vec`.

I missed it because I checked all uses of `from_vec`, not the actual call stack, and this use in locally typical calls `from_iter` which calls `from_vec` with  truncated `Candidates`, causing the crash.

```rust
        indices
            .iter()
            .take(new_len)
            .map(|&index| new.data[index])
            .collect()
```

The model will now have to be updated, but whatever.